requirements.txt
streamlit
pandas
numpy
matplotlib.pyplot
sklearn.preprocessing
keras.models
keras.layers.convolutional
keras.layers
keras.backend
base64
sklearn.metrics
math

